Fix the blank query issue retrieving all email. Modularise the code into separate files.
